+++ drivers/staging/android/ion/ion_heap.c
@@ -117,132 +117,40 @@ int ion_heap_map_user(struct ion_heap *heap, struct ion_buffer *buffer,
 	return 0;
 }
 
-#define MAX_VMAP_RETRIES 10
-
-/**
- * An optimized page-zero'ing function. vmaps arrays of pages in large
- * chunks to minimize the number of memsets and vmaps/vunmaps.
- *
- * Note that the `pages' array should be composed of all 4K pages.
- *
- * NOTE: This function does not guarantee synchronization of the caches
- * and thus caller is responsible for handling any cache maintenance
- * operations needed.
- */
-int ion_heap_pages_zero(struct page **pages, int num_pages)
-{
-	int i, j, npages_to_vmap;
-	void *ptr = NULL;
-
-	/*
-	 * As an optimization, we manually zero out all of the pages
-	 * in one fell swoop here. To safeguard against insufficient
-	 * vmalloc space, we only vmap `npages_to_vmap' at a time,
-	 * starting with a conservative estimate of 1/8 of the total
-	 * number of vmalloc pages available.
-	 */
-	npages_to_vmap = ((VMALLOC_END - VMALLOC_START)/8)
-			>> PAGE_SHIFT;
-	for (i = 0; i < num_pages; i += npages_to_vmap) {
-		npages_to_vmap = min(npages_to_vmap, num_pages - i);
-		for (j = 0; j < MAX_VMAP_RETRIES && npages_to_vmap;
-			++j) {
-			ptr = vmap(&pages[i], npages_to_vmap,
-					VM_IOREMAP, PAGE_KERNEL);
-			if (ptr)
-				break;
-			else
-				npages_to_vmap >>= 1;
-		}
-		if (!ptr)
-			return -ENOMEM;
-
-		memset(ptr, 0, npages_to_vmap * PAGE_SIZE);
-		vunmap(ptr);
-	}
-
-	return 0;
-}
-
-int ion_heap_alloc_pages_mem(struct pages_mem *pages_mem)
-{
-	struct page **pages;
-	unsigned int page_tbl_size;
-
-	pages_mem->free_fn = kfree;
-	page_tbl_size = sizeof(struct page *) * (pages_mem->size >> PAGE_SHIFT);
-	if (page_tbl_size > SZ_8K) {
-		/*
-		 * Do fallback to ensure we have a balance between
-		 * performance and availability.
-		 */
-		pages = kmalloc(page_tbl_size,
-				__GFP_COMP | __GFP_NORETRY |
-				__GFP_NO_KSWAPD | __GFP_NOWARN);
-		if (!pages) {
-			pages = vmalloc(page_tbl_size);
-			pages_mem->free_fn = vfree;
-		}
-	} else {
-		pages = kmalloc(page_tbl_size, GFP_KERNEL);
-	}
-
-	if (!pages)
-		return -ENOMEM;
-
-	pages_mem->pages = pages;
-	return 0;
-}
-
-void ion_heap_free_pages_mem(struct pages_mem *pages_mem)
-{
-	pages_mem->free_fn(pages_mem->pages);
-}
-
-int ion_heap_high_order_page_zero(struct page *page, int order)
-{
-	int i, ret;
-	struct pages_mem pages_mem;
-	int npages = 1 << order;
-	pages_mem.size = npages * PAGE_SIZE;
-
-	if (ion_heap_alloc_pages_mem(&pages_mem))
-		return -ENOMEM;
-
-	for (i = 0; i < (1 << order); ++i)
-		pages_mem.pages[i] = page + i;
-
-	ret = ion_heap_pages_zero(pages_mem.pages, npages);
-	dma_sync_single_for_device(NULL, page_to_phys(page), pages_mem.size,
-				   DMA_BIDIRECTIONAL);
-	ion_heap_free_pages_mem(&pages_mem);
-	return ret;
-}
-
 int ion_heap_buffer_zero(struct ion_buffer *buffer)
 {
 	struct sg_table *table = buffer->sg_table;
+	pgprot_t pgprot;
 	struct scatterlist *sg;
-	int i, j, ret = 0, npages = 0;
-	struct pages_mem pages_mem;
+	struct vm_struct *vm_struct;
+	int i, j, ret = 0;
 
-	pages_mem.size = PAGE_ALIGN(buffer->size);
+	if (buffer->flags & ION_FLAG_CACHED)
+		pgprot = PAGE_KERNEL;
+	else
+		pgprot = pgprot_writecombine(PAGE_KERNEL);
 
-	if (ion_heap_alloc_pages_mem(&pages_mem))
+	vm_struct = get_vm_area(PAGE_SIZE, VM_ALLOC);
+	if (!vm_struct)
 		return -ENOMEM;
 
 	for_each_sg(table->sgl, sg, table->nents, i) {
 		struct page *page = sg_page(sg);
 		unsigned long len = sg_dma_len(sg);
 
-		for (j = 0; j < len / PAGE_SIZE; j++)
-			pages_mem.pages[npages++] = page + j;
+		for (j = 0; j < len / PAGE_SIZE; j++) {
+			struct page *sub_page = page + j;
+			struct page **pages = &sub_page;
+			ret = map_vm_area(vm_struct, pgprot, &pages);
+			if (ret)
+				goto end;
+			memset(vm_struct->addr, 0, PAGE_SIZE);
+			unmap_kernel_range((unsigned long)vm_struct->addr,
+					   PAGE_SIZE);
+		}
 	}
-
-	ret = ion_heap_pages_zero(pages_mem.pages, npages);
-	dma_sync_sg_for_device(NULL, table->sgl, table->nents,
-			       DMA_BIDIRECTIONAL);
-	ion_heap_free_pages_mem(&pages_mem);
+end:
+	free_vm_area(vm_struct);
 	return ret;
 }
