+++ drivers/staging/android/ion/ion_page_pool.c
@@ -35,7 +35,7 @@ static void *ion_page_pool_alloc_pages(struct ion_page_pool *pool)
 		return NULL;
 
 	if (pool->gfp_mask & __GFP_ZERO)
-		if (ion_heap_high_order_page_zero(page, pool->order))
+		if (msm_ion_heap_high_order_page_zero(page, pool->order))
 			goto error_free_pages;
 
 	return page;
